pyproject.toml
[project]
name = "XJTUToolBox"
version = "1.2.3"
requires-python = ">=3.10, <=3.13"
readme = "README.md"
description = "A desktop application offering a variety of services for Xi'an Jiaotong University, including attendance check, schedule inquiry, and one-key teaching evaluation."
dependencies = [
    "click>=8.1.0",
    "requests>=2.33.0",
    "fake-useragent~=2.2.0",
    "pycryptodome~=3.21.0",
    # 【环境隔离】为 macOS, Windows, Linux x86_64 安装 UI 库（跳过 Linux aarch64）
    "pyqt-fluent-widgets~=1.8.7; sys_platform == 'darwin' or sys_platform == 'win32' or (sys_platform == 'linux' and platform_machine == 'x86_64')",
    "pyqt5-frameless-window~=0.7.4; sys_platform == 'darwin' or sys_platform == 'win32' or (sys_platform == 'linux' and platform_machine == 'x86_64')",
    # 【环境隔离】为 macOS, Linux x86_64 安装核心 PyQt5（跳过 Linux aarch64）
    "pyqt5>=5.15.11; sys_platform == 'darwin' or (sys_platform == 'linux' and platform_machine == 'x86_64')",
    # 在 Windows 下，强制使用 pyqt5>=5.15.11，配合软件包 pyqt5-qt5==5.15.2
    "pyqt5>=5.15.11; sys_platform == 'win32'",
    "pyqt5-qt5>=5.15,<=5.15.2; sys_platform == 'win32'",
    "platformdirs~=4.3.6",
    "peewee~=3.17.8",
    "icalendar~=6.1.0",
    "packaging~=24.2",
    "markdown~=3.7",
    "colorama~=0.4.6",
    "psutil~=7.0.0",
    "tqdm~=4.67.1",
    "lxml~=5.3.1",
    "plyer~=2.1.0",
    "pytz~=2025.2",
    "pyjwt~=2.12.0",
    "keyring~=25.6.0",
    "Cython~=3.2.1; sys_platform == 'darwin'",
    "pyobjus>=1.2.4; sys_platform == 'darwin'",
    "urllib3>=2.6.3",
    "pillow>=12.2.0",
]

[tool.uv]
package = false
# 支持平台：x86/x86_64 Windows, x86_64/arm64 macOS, x86_64/aarch64 linux
required-environments = [
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
    "sys_platform == 'win32' and platform_machine == 'x86'",
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
    # 【注册平台】必须注册 Linux aarch64，否则 uv lock 会报错
    "sys_platform == 'linux' and platform_machine == 'aarch64'",
]

[dependency-groups]
dev = [
    "pillow>=12.0.0",
    "pyinstaller>=6.17.0",
    "pyqt5-stubs>=5.15"
]

[project.scripts]
xjtutoolbox = "cli.main:cli"
xjtutool = "cli.main:cli"
